+++ RELEASE_NOTES.md
@@ -5,6 +5,8 @@ A new header is inserted each time a *tag* is created.
 
 ## main branch
 
+- backend: streamline texture upload APIs [⚠️ **API Change**]
+
 ## v1.25.3
 
 - engine: Fix Adreno gpu crash introduced by gpu morph target change